G/MM: Don't use 64-bit integers for sector times
32-bit unsigned ones are fine and that's what
the actual script code uses.

+++ source/game/g_ascript.cpp
@@ -1781,7 +1781,7 @@ static void objectGameClient_NewRaceRun( int numSectors, gclient_t *self ) {
 	StatsowFacade::Instance()->NewRaceRun( PLAYERENT( playerNum ), numSectors );
 }
 
-static void objectGameClient_SetSectorTime( int sector, int64_t time, gclient_t *self ) {
+static void objectGameClient_SetSectorTime( int sector, uint32_t time, gclient_t *self ) {
 	// TODO: Validate `self`
 	int playerNum = objectGameClient_PlayerNum( self );
 	// TODO: Throw a script exception at this!
@@ -1792,7 +1792,7 @@ static void objectGameClient_SetSectorTime( int sector, int64_t time, gclient_t 
 	StatsowFacade::Instance()->SetSectorTime( PLAYERENT( playerNum ), sector, time );
 }
 
-static RunStatusQuery *objectGameClient_CompleteRaceRun( int64_t finalTime, gclient_t *self ) {
+static RunStatusQuery *objectGameClient_CompleteRaceRun( uint32_t finalTime, gclient_t *self ) {
 	int playerNum = objectGameClient_PlayerNum( self );
 	// TODO: Throw a script exception at this!
 	if( playerNum < 0 || playerNum >= gs.maxclients ) {
@@ -1883,8 +1883,8 @@ static const asMethod_t gameclient_Methods[] =
 	{ ASLIB_FUNCTION_DECL( void, set_chaseActive, ( const bool active ) ), asFUNCTION( objectGameClient_SetChaseActive ), asCALL_CDECL_OBJLAST },
 	{ ASLIB_FUNCTION_DECL( bool, get_chaseActive, ( ) const ), asFUNCTION( objectGameClient_GetChaseActive ), asCALL_CDECL_OBJLAST },
 	{ ASLIB_FUNCTION_DECL( void, newRaceRun, ( int numSectors ) ), asFUNCTION( objectGameClient_NewRaceRun ), asCALL_CDECL_OBJLAST },
-	{ ASLIB_FUNCTION_DECL( void, setSectorTime, ( int sector, int64 time ) ), asFUNCTION( objectGameClient_SetSectorTime ), asCALL_CDECL_OBJLAST },
-	{ ASLIB_FUNCTION_DECL( RunStatusQuery @, completeRaceRun, ( int64 finalTime ) ), asFUNCTION( objectGameClient_CompleteRaceRun ), asCALL_CDECL_OBJLAST },
+	{ ASLIB_FUNCTION_DECL( void, setSectorTime, ( int sector, uint time ) ), asFUNCTION( objectGameClient_SetSectorTime ), asCALL_CDECL_OBJLAST },
+	{ ASLIB_FUNCTION_DECL( RunStatusQuery @, completeRaceRun, ( uint finalTime ) ), asFUNCTION( objectGameClient_CompleteRaceRun ), asCALL_CDECL_OBJLAST },
 	{ ASLIB_FUNCTION_DECL( void, setHelpMessage, ( uint msg ) ), asFUNCTION( objectGameClient_SetHelpMessage ), asCALL_CDECL_OBJLAST },
 	{ ASLIB_FUNCTION_DECL( void, setQuickMenuItems, ( const String &in ) ), asFUNCTION( objectGameClient_SetQuickMenuItems ), asCALL_CDECL_OBJLAST },
